app/requirements.txt
# Web scraping
newspaper3k==0.2.8
trafilatura==1.6.0
beautifulsoup4==4.12.2

# GenAI
langchain==0.0.267
openai==0.28.0
tiktoken==0.5.1

# Vector Database
chromadb==0.4.13
qdrant-client==1.4.0

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Utilities
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.0
pytest==7.4.2
